rename discriminator & AE->VAE, and merge code in VAE & VQVAE
